Refactored servicebus registration
