+++ src/Dfe.SignIn.Web.SelectOrganisation/Program.cs
@@ -77,7 +77,7 @@ builder.Services
     .SetupSelectOrganisationInteractions();
 
 builder.Services
-    .AddServiceBusIntegration(builder.Configuration, azureTokenCredential);
+    .AddServiceBusClient(builder.Configuration, azureTokenCredential);
 
 if (builder.Environment.IsEnvironment("Local")) {
     builder.Services.AddNullInteractor<WriteToAuditRequest, WriteToAuditResponse>();
